softmax reward model outputs
Logits appear to be the outputs of both open_assistant and reciprocate models, but then requires softmax treatment for correct interpretation. Linear normalization via EMA tracking can then be removed, due to softmax itself just having a relative interpretation and always summing to 1. This change is intended to improve the interpretation of these reward model outputs.